refactor(core): remove duplicated code in change_detection_util
Removes duplicated code in change_detection_util and reorganizes imports to use utils module.
