+++ packages/core/src/change_detection/change_detection.ts
@@ -12,7 +12,7 @@ import {IterableDifferFactory, IterableDiffers} from './differs/iterable_differs
 import {KeyValueDifferFactory, KeyValueDiffers} from './differs/keyvalue_differs';
 
 export {SimpleChange, SimpleChanges} from '../interface/simple_change';
-export {devModeEqual} from './change_detection_util';
+export {devModeEqual} from '../util/comparison';
 export {ChangeDetectorRef} from './change_detector_ref';
 export {ChangeDetectionStrategy, ChangeDetectorStatus, isDefaultChangeDetectionStrategy} from './constants';
 export {DefaultIterableDiffer, DefaultIterableDifferFactory} from './differs/default_iterable_differ';
