+++ packages/core/src/change_detection/differs/default_iterable_differ.ts
@@ -7,8 +7,8 @@
  */
 
 import {RuntimeError, RuntimeErrorCode} from '../../errors';
+import {isListLikeIterable, iterateListLike} from '../../util/iterable';
 import {stringify} from '../../util/stringify';
-import {isListLikeIterable, iterateListLike} from '../change_detection_util';
 
 import {IterableChangeRecord, IterableChanges, IterableDiffer, IterableDifferFactory, NgIterable, TrackByFunction} from './iterable_differs';
