+++ packages/core/test/linker/query_list_spec.ts
@@ -7,8 +7,8 @@
  */
 
 import {ɵgetDOM as getDOM} from '@angular/common';
-import {iterateListLike} from '@angular/core/src/change_detection/change_detection_util';
 import {QueryList} from '@angular/core/src/linker/query_list';
+import {iterateListLike} from '@angular/core/src/util/iterable';
 import {fakeAsync, tick} from '@angular/core/testing';
 
 {
